Add files via upload
https://github.ibm.com/wdp-gov/catalog-api/blob/dev/catalog-scripts/FixOrphanedCatalog.md

Fix an orphaned catalog
In a CPD environment where we have an orphaned catalog, here are the steps to add another user to that catalog as admin.

If there exists a catalog where there is no administrator still in the user management system allowing for access to that catalog, below steps detail how to repair that by adding a new user as admin for the catalog. This can be done via the POST /v2/catalogs/{catalog_id}/members api call (provided a script -- details below).

 To make this api call, the following pieces of information are needed: -- current uid of the desired new administrator user (see step 1) -- serviceId token (see step 2) -- catalog_id of the inaccessible catalog (see step 3)

Scripts are provided for each step. However, in each script, there are pieces of information that you will need to fill out, like the cluster address, or the user password.